+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Add "Admin Settings" tab to the Updater/Admin Panel
     * Add toggleable option to persist system logs to Admin Settings
     * Move "Set password" page to be accessible via Admin Settings tab
+  * Add success/fail messages to zones settings changes
 * System
   * Make update process properly report errors